Move GCManagementFactory out of GCImpl.
